[mod_tile] Fix get_load_avg calculation
mod_tile.c was missing an include file resulting in a wrong return type

+++ mod_tile.c
@@ -50,7 +50,7 @@ module AP_MODULE_DECLARE_DATA tile_module;
 #include "store.h"
 #include "dir_utils.h"
 #include "mod_tile.h"
-
+#include "sys_utils.h"
 
 
 #if !defined(OS2) && !defined(WIN32) && !defined(BEOS) && !defined(NETWARE)
@@ -696,7 +696,7 @@ static int tile_handler_dirty(request_rec *r)
 static int tile_storage_hook(request_rec *r)
 {
 //    char abs_path[PATH_MAX];
-    int avg;
+    double avg;
     int renderPrio = 0;
     enum tileState state;
 
@@ -760,7 +760,7 @@ should already be done
             } else if (avg > scfg->max_load_old) {
                // Too much load to render it now, mark dirty but return old tile
                request_tile(r, cmd, 0);
-               ap_log_rerror(APLOG_MARK, APLOG_DEBUG, 0, r, "Load (%d) larger max_load_old (%d). Mark dirty and deliver from cache.", avg, scfg->max_load_old);
+               ap_log_rerror(APLOG_MARK, APLOG_DEBUG, 0, r, "Load (%f) larger max_load_old (%d). Mark dirty and deliver from cache.", avg, scfg->max_load_old);
                if (!incFreshCounter(OLD, r)) {
                    ap_log_rerror(APLOG_MARK, APLOG_WARNING, 0, r,
                         "Failed to increase fresh stats counter");
@@ -772,7 +772,7 @@ should already be done
         case tileMissing:
             if (avg > scfg->max_load_missing) {
                request_tile(r, cmd, 0);
-               ap_log_rerror(APLOG_MARK, APLOG_INFO, 0, r, "Load (%d) larger max_load_missing (%d). Return HTTP_NOT_FOUND.", avg, scfg->max_load_missing);
+               ap_log_rerror(APLOG_MARK, APLOG_INFO, 0, r, "Load (%f) larger max_load_missing (%d). Return HTTP_NOT_FOUND.", avg, scfg->max_load_missing);
                if (!incRespCounter(HTTP_NOT_FOUND, r, cmd, rdata->layerNumber)) {
                    ap_log_rerror(APLOG_MARK, APLOG_WARNING, 0, r,
                         "Failed to increase response stats counter");
